+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Support the ArchLinux Node.js 17 ABI - see https://github.com/nodejs/node/issues/41410
  - Fix #25, always close open Datasets when the process exits
  - `gdal.calcAsync` now correctly returns a rejected Promise instead of triggering an uncatchable error when handling an exception in the user-supplied transformation function
+ - Fix #26, some methods do not check the passed object class
 
 ## [3.4.1] 2021-01-05